refactor: change product classes structure by BOUNDED CONTEXT
